parser cleanup + test cases
