+++ test/test_parser.py
@@ -1,26 +1,26 @@
 """Testing the database in a general sense and the models specifically."""
 
-#from sqlalchemy.exc import IntegrityError, ProgrammingError
-
-from . import KaosTestCase
-#from .context import kaos
+from sqlalchemy.exc import IntegrityError, ProgrammingError
+from sqlalchemy.sql import operators
 
 from . import KaosTestCase
 from .context import kaos
-from kaos.models import DB, SatelliteInfo
-from kaos.ephemeris_parser import EphemerisParser
-
+from kaos.models import *
+from kaos.parser import *
 
 class TestEphemerisParser(KaosTestCase):
     """Ensures that the response history table behaves as expected."""
 
-    def test_simple_parser1(self):
-        """Test that an empty response cannot be commited to history."""
-        sat = SatelliteInfo( platform_name="TEST HOMIE" )
-        sat.save()
-        DB.session.commit()
-        EphemerisParser.parse_file("/home/zeyadtamimi/projects/cpen491/kaos/ephemeris/Radarsat2_J2000.e", 1)
+    def test_simple_parser_J2000(self):
+        parse_ephemeris_file("ephemeris/Radarsat2_J2000.e")
+        orbit = OrbitRecords()
 
-    #def test_simple_parser2(self):
-    #    """Test that an empty response cannot be commited to history."""
-    #    EphemerisParser.parse_file("/Users/jamieasefa/Documents/Courses/CPEN491/kaos/ephemeris/Radarsat2_Fixed.e", 1)
+        # test that the correct number of entries was created
+        self.assertTrue(len(orbit.query.all()) == 17307) #taken from ephem file
+
+    def test_simple_parser_Fixed(self):
+        parse_ephemeris_file("ephemeris/Radarsat2_Fixed.e")
+        orbit = OrbitRecords()
+
+        # test that the correct number of entries was created
+        self.assertTrue(len(orbit.query.all()) == 17307) #taken from ephem file
